% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{aa_proportions}
\alias{aa_proportions}
\title{Plotting of the porportions of amino acids in a sequence}
\usage{
aa_proportions(aa_seq)
}
\arguments{
\item{aa_seq}{Given sequence of amino acids}
}
\value{
proportions, a plot of counts of each amino acid in the sequence
}
\description{
Plotting of the porportions of amino acids in a sequence
}
\examples{
aa_proportions("MVAAA")
}
