% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_aa.R
\name{codon_to_aa}
\alias{codon_to_aa}
\title{Codon to amino acid}
\usage{
codon_to_aa(codon)
}
\arguments{
\item{codon}{sequence of 3 nucleotides}
}
\value{
aa, the amino acid corresponding to the given codon
}
\description{
Codon to amino acid
}
\examples{
codon_to_aa("AUG")
}
