% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_DNA_string.R
\name{create_DNA_string}
\alias{create_DNA_string}
\title{Title: Create a random DNA string}
\usage{
create_DNA_string(length_of_DNA_string)
}
\arguments{
\item{length_of_DNA_string}{Length of returned DNA string}
}
\value{
a DNA string as a character with the length of the input parameter
}
\description{
Title: Create a random DNA string
}
\examples{
create_DNA_string(10)
}
