% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{DNA_to_RNA}
\alias{DNA_to_RNA}
\title{Title: DNA sequences to RNA sequences}
\usage{
DNA_to_RNA(DNA_string)
}
\arguments{
\item{DNA_string}{A DNA sequence to be changed to an RNA sequence}
}
\value{
Returns the corresponding RNA sequence to the input DNA sequence
}
\description{
Title: DNA sequences to RNA sequences
}
