% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_codon.R
\name{extract_char}
\alias{extract_char}
\title{Extract and Plot Character Counts}
\usage{
extract_char(split_char)
}
\arguments{
\item{split_char}{A character string to analyze.}
}
\value{
A ggplot2 bar plot displaying unique characters and their counts.
}
\description{
This function extracts unique characters from a given string, counts their occurrences,
and generates a bar plot of these counts.
}
\examples{
# Example usage of extract_char
sample_string <- "ABCAABB"
extract_char(sample_string)

}
