% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Creating codons}
\usage{
extract_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{Clarifies that this input is a DNA/RNA sequence}

\item{start}{An argument specifying the starting position in the sequence to extract codons}
}
\value{
This code defines a function, base_pairs, that takes a DNA or RNA sequence (sequence_length) and divides it into codons
}
\description{
Creating codons
}
