% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Random sequence of bases}
\usage{
function1(name_me2)
}
\arguments{
\item{name_m2}{is a number defining the length of the sequence}
}
\value{
Creates a random sequence of bases
}
\description{
Random sequence of bases
}
\examples{
function(10)
}
