% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{function2}
\alias{function2}
\title{Convert from DNA to RNA}
\usage{
function2(name_me2)
}
\arguments{
\item{name_me2}{DNA sequence}
}
\value{
Converts from DNA to RNA
}
\description{
Convert from DNA to RNA
}
\examples{
(function2("AATTGGAT"))
}
