% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{function3}
\alias{function3}
\title{Divides sequence into codons}
\usage{
function3(name_me2, start = 1)
}
\arguments{
\item{name_me2}{DNA/RNA sequnce}

\item{start}{Defines reading frame}
}
\value{
Divides sequence into codons
}
\description{
Divides sequence into codons
}
\examples{
(function3("AATTTGGAA",1))
}
