% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4.R
\name{function4}
\alias{function4}
\title{Codon translator}
\usage{
function4(codons)
}
\arguments{
\item{codons}{A sequence of codons}
}
\value{
Translates DNA sequence to amino acids
}
\description{
Codon translator
}
\examples{
function4("AAG","GGA","AGT")
}
