% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function5.R
\name{function5}
\alias{function5}
\title{AA frequency plotter}
\usage{
function5(name_me2)
}
\arguments{
\item{name_me2}{Amino acid sequence}
}
\value{
Bar plot showing frequency of each amino acid
}
\description{
AA frequency plotter
}
\examples{
(function5("ILKA"))
}
