% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codon.R
\name{generate_codon}
\alias{generate_codon}
\title{Generate a codon}
\usage{
generate_codon(gen_seq, start = 1)
}
\arguments{
\item{gen_seq}{enter the genome to create codon}

\item{start}{the start character for codon creation}
}
\value{
the function will generate partitions of codons based on the gene that has inserted.
}
\description{
Generate a codon
}
\examples{
generate_codon("GTCGTC")
}
