% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rand_DNA.R
\name{generate_rand_DNA}
\alias{generate_rand_DNA}
\title{Generate_rand_DNA}
\usage{
generate_rand_DNA(DNA_len)
}
\arguments{
\item{DNA_len}{wanted length of the random generated DNA}
}
\value{
The function will return a random DNA sequence of the desired length
}
\description{
Generate_rand_DNA
}
\examples{
generate_rand_DNA(3)
}
