% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_name.R
\name{translate_codons}
\alias{translate_codons}
\title{Translate codon sequence to amino acids}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{translate codon sequence to amino acids}
}
\value{
function output is an amino acid sequence
}
\description{
Translate codon sequence to amino acids
}
