% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_rna.R
\name{dna_to_rna}
\alias{dna_to_rna}
\title{DNA_to_RNA}
\usage{
dna_to_rna(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{Sequence of the nucleotides A, C, G and T}
}
\value{
Sequence of the nucleotides A, C, G and U corresponding to the DNA sequence given as parameter
}
\description{
DNA_to_RNA
}
\examples{
dna_to_rna("A C G T")
}
